+++ CHANGELOG.md
@@ -7,6 +7,18 @@ Version 1.2-SNAPSHOT
 ADDED:
 
 - new bean class ComponentPositionModel for store the position of a component
+- new class ComponentPositionStore for store the position of a component in the java preferences
+- new enum class ComponentPositionPreference that holds the key suffixes that will be saved in the preferences
+
+CHANGED:
+
+- update gradle to new version 8.5
+- update of dependency lombok version to 1.18.30
+- update of gradle-plugin dependency 'io.freefair.gradle:lombok-plugin' to new version 8.4
+- update of gradle-plugin dependency 'com.github.ben-manes.versions.gradle.plugin' to new version 0.50.0
+- update of gradle-plugin dependency 'org.ajoberstar.grgit:grgit-gradle' to new version 5.2.1
+- update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new version 6.23.3
+- update of test dependency junit-jupiter version to new patch version 5.10.1
 
 Version 1.1
 -------------